Removed yat api, integrated Yat lib on wallet creation flow
