--- app/src/main/java/com/tari/android/wallet/infrastructure/yat/emojiid/YatEmojiIdAPI.kt
@@ -1,55 +1,0 @@
-/**
- * Copyright 2020 The Tari Project
- *
- * Redistribution and use in source and binary forms, with or
- * without modification, are permitted provided that the
- * following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above
- * copyright notice, this list of conditions and the following disclaimer in the
- * documentation and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of
- * its contributors may be used to endorse or promote products
- * derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
- * CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
- * INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
- * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
- * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
- * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package com.tari.android.wallet.infrastructure.yat.emojiid
-
-import com.tari.android.wallet.BuildConfig
-import retrofit2.Call
-import retrofit2.http.*
-
-interface YatEmojiIdAPI {
-
-    @GET("emoji")
-    // this header value is temporary for the scratch-based API deployment
-    @Headers("Authorization: Basic " + BuildConfig.YAT_API_BASIC_AUTH_KEY)
-    fun getSupportedSet(): Call<List<String>>
-
-    @GET("emoji_id/search")
-    fun check(@Query("emoji_id") emojiId: String): Call<EmojiIdCheckResponseBody>
-
-    @PATCH("emoji_id/{emoji_id}")
-    fun edit(@Path("emoji_id") emojiId: String, @Body body: EmojiIdEditRequestBody): Call<Void>
-
-    @GET("emoji_id/{emoji_id}")
-    fun search(@Path("emoji_id") emojiId: String): Call<EmojiIdSearchResponseBody>
-
-}
