+++ yat-example.properties
@@ -1,0 +1,9 @@
+# This file must *NOT* be checked into Version Control Systems,
+# as it contains sensitive app configuration data.
+
+yat.appName=YOUR_APPNAME
+yat.sourceName=YOUR_SOURCE_NAME
+yat.pathKey=YOUR_PARTNER_NAME
+yat.pubKey=YOUR_PUBLIC_KEY
+yat.code=YOUR_CODE
+yat.authToken=YOUR_TOKEN
